Use compiler endpoint artifacts for GraphQL endpoint export
