+++ compiler-plugin/build.gradle
@@ -35,8 +35,6 @@ dependencies {
 
     implementation project(':graphql-commons')
 
-    implementation "com.fasterxml.jackson.core:jackson-databind:${jacksonVersion}"
-    implementation "com.fasterxml.jackson.dataformat:jackson-dataformat-yaml:${jacksonVersion}"
     implementation "org.apache.commons:commons-text:1.12.0"
 }
