+++ compiler-plugin/src/main/java/module-info.java
@@ -21,8 +21,6 @@ module io.ballerina.stdlib.graphql.compiler {
     requires io.ballerina.tools.api;
     requires io.ballerina.parser;
     requires io.ballerina.stdlib.graphql.commons;
-    requires com.fasterxml.jackson.dataformat.yaml;
-    requires com.fasterxml.jackson.databind;
     requires org.apache.commons.text;
 
     exports io.ballerina.stdlib.graphql.compiler;
